requirements.txt
SQLAlchemy
psycopg2-binary
asyncpg
python-decouple
fastapi
sqlalchemy==2.0.3
uvicorn
tenacity
redis
rq
